Add the MainMoviesComponent
