+++ src/components/abstract-component.js
@@ -24,7 +24,7 @@ export default class AbstractComponent {
   }
 
   removeElement() {
-    this._element.remove();
+    this.getElement().remove();
     this._element = null;
   }
 }
