+++ src/components/movies-container.js
@@ -3,11 +3,6 @@ import AbstractComponent from './abstract-component';
 const createMoviesContainerMarkup = () => {
   return (
     `<section class="films">
-      <section class="films-list">
-        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-        <div class="films-list__container"></div>
-        <button class="films-list__show-more">Show more</button>
-      </section>
     </div>`
   );
 };
@@ -15,23 +10,5 @@ const createMoviesContainerMarkup = () => {
 export default class MoviesContainer extends AbstractComponent {
   getTemplate() {
     return createMoviesContainerMarkup();
-  }
-
-  getMoviesListElement() {
-    return this.getElement().querySelector(`.films-list__container`);
-  }
-
-  setLoadButtonCallback(callback) {
-    this.getElement().querySelector(`.films-list__show-more`)
-      .addEventListener(`click`, callback);
-  }
-
-  toggleShowLoadButton(areAllMoviesShown) {
-    const loadMoreButton = this.getElement().querySelector(`.films-list__show-more`);
-    if (areAllMoviesShown) {
-      loadMoreButton.style.display = `none`;
-    } else {
-      loadMoreButton.style.display = ``;
-    }
   }
 }
